making changes to consolidate class view
